Invert name of Listing and Page panel
